requirements.txt
streamlit>=1.34.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
PyPDF2>=3.0.1
transformers>=4.41.0
torch>=2.2.0
fpdf>=1.7.2
numpy>=1.24.0
tqdm>=4.66.0
